fix executionPoker
